M1579 Refactoring cleanup
